backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend of the Dadly application, which is a website for finding and cooking recipes with swipe functionality."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "bcrypt>=4.0.0",
    "cryptography>=43.0.0",
    "fastapi>=0.119.1",
    "git-filter-repo>=2.47.0",
    "httpx>=0.27.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.12.3",
    "pymysql>=1.1.2",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
    "-v"
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests"
]
